requirements.txt
openai
requests
replicate
opencv-python
numpy
transformers
torch
Pillow